requirements.txt
streamlit
langchain
langchain-community
chromadb
ollama
python-dotenv
requests
